Add and integrate CPU tick
This is needlessly complex but was a great way to try out some design
patterns in Rust. The design uses the visitor pattern to dispatch CPU
instructions to the Sm83, which avoid a massive match statement, it uses
abstract factory to create opcode decoders, which allows us to keep
opcodes modular and separate how they are decoded from binary to how
they are serialized, the composite pattern is used to aggregate all of
the opcodes into a "master" decoder, and there's probably others I had
used that I can't recall off the top of my head.

Anyway, more complex than it needs to be, but was a fun way to learn
some Rust stuff :)

May revisit this to make it more functional rather than OO at some point
in time.