Refactor opcodes
Reorganized code such that there is a submodule for each opcode
mneumonic, which contains implementations for its associated traits.

Also separated the Cpu tick API from the instruction set into separate
traits. This allows the client to simply use the Cpu trait to drive the
Cpu logic.

todo: It is becoming obvious that the instruciton implementations in
sm83 isn't going to scale well. Need to figure out how to modularize
this a bit more. I tried separating the traits out, but this breaks
encapsulation since struct fields need to become public.